Added a "\n" at the end of every write
The method which writes the files file does not terminate
with a space or a carriage return. This causes ABINIT to hang
while reading the list of pseudopotentials.

Adding a "\n" to every write fixes this issue, but is not
elegant, as it adds useless "\n" everywhere, not only in the
files file.  A better solution may be desirable?