Added a "\n" at the end of every write
The method which writes the files file does not terminate
with a space or a carriage return. This causes ABINIT to hang
while reading the list of pseudopotentials.

Adding a "\n" to every write fixes this issue, but is not
elegant, as it adds useless "\n" everywhere, not only in the
files file.  A better solution may be desirable?


Former-commit-id: caa737b15bc5b589f3a4cb21a683700c083a6309 [formerly 33881a4a9c9394f3b90aa0fc32565f0f18f5f28f]
Former-commit-id: 31319227859d34327ffe93eaa50be7a50f62b67f