Make it easier for user to submit custom LDAUU
Currently, it's non-obvious how a user can add a custom `LDAUU` value to their `user_incar_settings` that would be used in the general case (regardless of what anions are in the input structure). This adds a fallback option, so that e.g. `user_incar_settings={'LDAUU':{'Fe':5}}` would work as expected.