Added get_symmetry_operations method
The SpacegroupAnalyzer class has a get_symmetry_operations() method
that returns a list of Symmop symmetry operations.
A PointGroupAnalyzer [sic] object has the list of relevant
symmetry operations stored in self.symmops.
This commit adds a PointGroupAnalyzer.get_symmetry_operations
method to give an API that follows the SpacegroupAnalyzer class.