Bugfix for PointGroupAnalyzer get_symmetry_operations
Commit 42df22fe0d2488c539914f58ed90961411eaec46 added a
get_symmetry_operations method to the PointGroupAnalyzer class.
It also introduced a bug, where the symmetry operations returned
only included those found from the initial analysis of a molecule,
instead of the complete set of symmetry operations for the
corresponding point group. This commit fixes this bug, and adds
a more detailed test that checks a full set of symmetry operations
is returned.