Alternate dielectric function parsing: vasp 5.4.4
VASP 5.4.4 calculates the frequency dependent dielectric response in two ways, using a "density-density" response function, and using a "velocity-velocity" response function (see the comments in linear_optics.F in the vasp source code). The latter is marked in OUTCAR as current-current (2nd set in vasprun.xml), but is not labelled in vasprun.xml.

This commit addresses an issue where the second unlabelled dielectric function would overwrite the first.

https://github.com/materialsproject/pymatgen/issues/940

For consistency with older versions of vasp, the "density-density" linear response function
is read into self.dielectric.
The "velocity-velocity" linear response function (described as "current-current" in OUTCAR)
is read into self.other_dielectric['current-current'].

This commit includes a test that if vasprun.xml contains a second unlabelled set of dielectric
repsonse data, this is read into self.other_dielectric['current-current'].

If a third unlablled set of dielectric data is present in the vasprun.xml, a NotImplementedError
is raise (also tested).