Bug fix for DiffusionAnalyzer.from_vasprun()
Commit c99dba04ebba648909ab5f71a7593f26cb388f11 introduces a bug
in the from_structures() method if this is called from the
from_vaspruns() method. This commit reverts this change.

In from_vasprun(), structures are collected as a generator function,
using get_structures(). The from_structures() method, however, expects
a list of structures, and tries to access structures[0]. This fails
if a constructor has been passed in, instead of a list (as is the
case with from_vasprun()).