Extend PWOutput implementation
I've written a large set of (working, this time) regular expressions for the PWOutput parser class, supporting the following properties:

- Energy
- Final energy
- Enthalpy (new enthalpy)
- Final enthalpy
- Density
- Warnings
- Total stress
- Stress tensor
- Total force
- Force on each atom with site and species data
- Bands data (kpoint coords, eigenvalues, occupations) **does not support spin-polarization**
- Fermi energy
- Convergence iterations
- Cell parameters
- Atomic positions
- PWSCF version
- Date
- Time
- Lattice type (bravais)
- Lattice parameter (alat)
- Unit cell volume
- Number of atoms
- Number of types
- Number of electrons
- Number of Kohn-Sham states
- KE cutoff (ecutwfc)
- Density cutoff (ecutrho)
- Electronic convergence threshold
- Mixing beta
- Number of iterations (used)
- Exchange-correlation
- celldm(n)
- lattice (a1, a2, a3)
- reciprocal lattice (b1, b2,  b3)
- Number of symmetry operations
- Number of kpoints
- Degauss
- Kpoints (cartesian, fractional) with index, coords, weight
- Job done

Planning to support:
- Symmetry operations
- anything else interesting that I've missed

The properties are listed in a patterns dictionary, as the are in the current PWOutput, but each entry is instead another dictionary containing the regular expression pattern, re flags, and the post-processing function for that property.

I've implemented a structures property that returns the structures in Structure objects, and I'm working on similar functions for band structure and symmetry operations.

I think that with all these properties, the current structure of PWOutput is unmanageable, but I'm not sure what a good alternative is. For the moment, I've continued to use the old structure.

As a relative beginner to this kind of programming, I'm interested in the process of designing an appropriate, maintainable implementation.