This commit tries to fix issue #1442 in the Waveder class.
The original issue is raised by the fact that the variables and member functions have the same name.
This has been fixed by prepending a underscore to the variable names.

Additionally, the datatype of the records read into cder should be of type np.complex64.
This has been changed.

Furthermore, the order of the reshape call for cder_data needs to be inverted, otherwise the mapping between bands, kpoints and spin won't be correct