build system: adopt PEP-518
The setup.py was using the deprecated `setup_requires` keyword, which
can result in dependencies being installed via `easy_install`.
`easy_install` is outdated and should no longer be used to install
packages.

This commit adopts PEP 518 and uses the `pyproject.toml` file to specify
the dependencies of the build system.