inputs.py: cp2k cannot auto-symmetrize the kpoint grid if SPGLIB is not available during runtime. I have added symmetry based kpoint reduction using the SpaceGroupAnalyzer to the from_kpoints() constructor. Also fixed a typo in E_DENSITY_GRID which didn't unpack the iterable.
outputs.py: when getting SCF opt, the HFX print-out after the preconditioner
and before the iterations was causing regrep to fail. I modified the
strip feature of the table reader to remove warnings and HFX info so this
doesn't happen. Other things might trip-up the SCF opt reader in the future
so this will be a good solution for general issues.

sets.py: updated the kpoints to reflect the new from_kpoints() constructor.
Also changed the default algo to standard if kpoints are present, which
for some reason CP2K requires.