A very subtle error... Everything in the sets/inputs preserves the ordering of the sites, except when you use the get_unique_site_indices function to get aliases such as {'Mg_1': [...], 'Mg_2': [...], 'O_1': [...]}.
Then, the sites are given by a alias: index dictionary that doesn't preserve
order, and so sites can get jumbled when this happens. It doesn't always happen, so
it was hard to identify. Now, when aliasing is used for the Coord section,
the keywords are created in order of their indices. To further protect against
these sorts of issues for unforseen problems, I'm changing "self.keywords"
to be stored as an OrderedDict once they are provided to the Section. Since
Ordered Dicts inherit from Dicts, this shouldn't change anything.

Lastly, I have applied black formatting to everything to keep consistent.