Add `run_metadata` to `required_bandfilling_params`
Add `run_metadata` to `required_bandfilling_params`, to check if the VASP calculation was a SOC calculation (vasp_ncl), which affects the determination of eigenvalue occupancies and hence the bandfilling correction. This parameter (`run_metadata`) is present and parsed by default when parsing defects with PyCDT (upon which these correction functions are based).