AqueousCompat: simplify treatment of H2O
After separating out the entropy correction for H2O, I
realized that by construction, the EnergyAdjustment
computed for water would always be zero. This makes
sense because the premise of this correction scheme
is to back-calculate the H2 energy, such that you
get the correct H2O formation energy from the DFT
energies of O2 and H2O.

I found it confusing to see a zero energy correction
applied to H2O, so this commit removes it. This
does not change the results of the class in any way.