Added functionality to 'get_symmetric_neighbor_list'. By setting unique=True each bond is only considered once (not double counted).
Previously each bond was essentially given twice in the output. If a bond from site i to j along R was in the list, the same bond in the inverted direction, namely from j to i along -R, was also in the output. By setting unique=True, only one of the bonds is given. The new functionality of 'are_symmetrically_related_bond' is used to keep track whether the bonds are reversed while being compared. This ensures that the symmetry operation actually generates the bond given in the list, not its reversed partner. This is important for building physical models (for example making sure that DM interaction has the right direction). Also, the inverted direction of a bond in a physical model corresponds to the hermitian conjugate, which is usually added manually, so there's no need to consider the 'full' neighbor list.