Relaxation Transformation
Sometimes, we have calculations with continue from a previous calculation. For example, you run a coarse relaxation followed by a fine relaxation in separate fireworks. In that case, it would be convenient to have a Transformation which represents that relaxation that can be part of the second calculations transformation history, rather than starting from the coarse-relaxed structure.

Perhaps this should be named something else, though? Relaxation is probably the most common potential use case, but it could be more general perhaps.