small bug in gruneisen.py
At converting self.acoustic_debye_temp from K to Hz we get numbers of the order 10^12 (Tera) but since w are not yet multiplied by 10^12, the value adt * const.value ("Boltzmann constant in Hz/K") should be divided by the const.tera. The error lead to that all values of w satisfy the condition while this should not be the case.