Added a verification method to PotcarSingle that attempts to use the sha256 hash that is present in new POTCAR files for checking file integrity. If the POTCAR file does not have such a hash in it, the md5 hash of the whole file is compared to the database of hashes.
If the comparison with the internal sha256 hash fails, that means
that the POTCAR is for sure corrupted and a ValueError is raised.
If the sha256 hash is not found, but the md5 hash of the file is
not in the database, only a warning is raised to conform to
previous behavior.