Redid BaseLammpsGenerator  Changes to be committed:
Reworked the way the input set generator generates the input file to be less cumbersome and use fewer if-else clauses, this new approach goes back to the original generator in order to be a lot more flexible (esp for custom lammps jobs). A LammpsSettings object is now present in order to validate the inputs specified by the user prior to actually passing them to lammps. This approach works out for the makers implemented currently (NVT/NPT/NVE/Minimization).